split app.go into handler files
